feat: Issue 107 (#8) Revised docker configuration - more configurable by all members
The new dockerfile + compose setup will allow using a postgresql container with ease in addition to the existing container.

If any changes are required to the docker container, they can easily be added into the dockerfile. This can also be tested in github actions to be published as a pre-built image for release.